Move in separated file struct MqttTestClient for helper tests
